Move WebSocketConnectionManager into its own file.
